add phone to api(ads)

+++ car_proj/database/migrations/2022_07_11_194511_create_dealership_hour_table.php
@@ -15,8 +15,8 @@ class CreateDealershipHourTable extends Migration
     {
         Schema::create('dealership_hour', function (Blueprint $table) {
             $table->id();
-            $table->integer('dealership_id')->unsigned();
-            $table->integer('hour_id')->unsigned();
+            $table->unsignedBigInteger('dealership_id')->unsigned();
+            $table->unsignedBigInteger('hour_id')->unsigned();
             $table->timestamps();
         });
     }
